Version 1.03.001 prepare for UAT

+++ WPI_CSA/WebService/WCEvent.swift
@@ -15,7 +15,7 @@ open class WCEvent {
     var startTime: Date
     var endTime: Date
     var location: String
-    var fee: Int?
+    var fee: Double?
     var ownerId: Int?
     var createdAt: Date?
     
@@ -59,7 +59,7 @@ open class WCEventManager {
                     event.ownerId = dict!["ownerId"]! as? Int
                     event.createdAt = (dict!["createdAt"]! as! String).Iso8601DateUTC
                     event.description = dict!["description"]! as! String
-                    if let fee = dict!["fee"] as? Int {
+                    if let fee = dict!["fee"] as? Double {
                         event.fee = fee
                     }
